requirements.txt
requests
urllib3
psycopg2-binary
coverage # para verificar os testes
